devshowcase_backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
requests==2.31.0
psycopg2-binary==2.9.9
python-slugify==8.0.1
django-ratelimit==4.1.0
PyJWT==2.8.0
cryptography==41.0.7
groq==1.0.0
python-dotenv==1.0.0
